+++ src/store/middleware/logoutMiddleware.ts
@@ -1,12 +1,19 @@
 import { Address } from '@multiversx/sdk-core/out';
 import { WritableDraft } from 'immer';
-import { storage } from 'storage';
-import { localStorageKeys } from 'storage/local';
-import { accountInfoSelector, isLoggedInSelector } from 'store/selectors';
+import {
+  addLoginExpiresAt,
+  removeLoginExpiresAt
+} from 'store/actions/loginInfo/loginInfoActions';
+import {
+  accountInfoSelector,
+  isLoggedInSelector,
+  loginExpiresAtSelector
+} from 'store/selectors';
 import { initialState as initialAccountState } from 'store/slices/account/accountSlice';
 import { initialState as initialLoginInfoState } from 'store/slices/loginInfo/loginInfoSlice';
 import { initialState as initialToastState } from 'store/slices/toast/toastSlice';
 import { initialState as initialTrackedTransactionsState } from 'store/slices/trackedTransactions/trackedTransactionsSlice';
+import { getStore } from 'store/store';
 import { StoreType } from '../store.types';
 
 export const resetStore = (store: WritableDraft<StoreType>) => {
@@ -22,20 +29,16 @@ export function getNewLoginExpiresTimestamp() {
 
 export function setLoginExpiresAt(expiresAt: number | null) {
   if (expiresAt == null) {
-    storage.local.removeItem(localStorageKeys.loginExpiresAt);
+    removeLoginExpiresAt();
     return;
   }
-  storage.local.setItem({
-    key: localStorageKeys.loginExpiresAt,
-    data: expiresAt,
-    expires: expiresAt
-  });
+  addLoginExpiresAt(expiresAt);
 }
 
 export const logoutMiddleware = (state: StoreType) => {
   const isLoggedIn = isLoggedInSelector(state);
+  const loginTimestamp = loginExpiresAtSelector(state);
   const { address, publicKey } = accountInfoSelector(state);
-  const loginTimestamp = storage.local.getItem(localStorageKeys.loginExpiresAt);
 
   if (address && publicKey !== new Address(address).hex()) {
     resetStore(state);
@@ -56,6 +59,6 @@ export const logoutMiddleware = (state: StoreType) => {
   if (isExpired) {
     // logout
     setLoginExpiresAt(null);
-    resetStore(state);
+    getStore().setState(resetStore);
   }
 };